Draw rooms
