+++ .classpath
@@ -1,7 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
 	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
+		<attributes>
+			<attribute name="module" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="lib" path="dependencies/java-json.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.base.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.controls.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.fxml.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.graphics.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.media.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.swing.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx.web.jar"/>
+	<classpathentry kind="lib" path="dependencies/javafx-swt.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
