+++ src/House/houseinput.json
@@ -1,0 +1,41 @@
+{
+  "Garage":{
+    "windows":1,
+    "lights":1,
+    "doorsTo":[
+      "Kitchen",
+    ]
+  },
+  "Kitchen":{
+    "windows":2,
+    "lights":3,
+    "doorsTo":[
+      "Garage",
+      "Living Room",
+      "Forklift Arena"
+    ]
+  },
+  "Living Room":{
+    "windows":2,
+    "lights":3,
+    "doorsTo":[
+      "Kitchen",
+      "Forklift Arena",
+    ]
+  },
+  "Forklift Arena":{
+    "windows":3,
+    "lights":8,
+    "doorsTo":[
+      "Living Room",
+      "Kitchen",
+      "Closet"
+    ]
+  },
+  "Closet":{
+    "windows":1,
+    "lights":1,
+    "doorsTo":["Forklift Arena"]
+  }
+
+}
