--- src/houseinput.json
@@ -1,16 +1,0 @@
-{
-  "Garage":{
-    "windows":0,
-    "lights":1,
-    "doorsTo":[
-      "Kitchen"
-    ]
-  },
-  "Kitchen":{
-    "windows":2,
-    "lights":2,
-    "doorsTo":[
-      "Garage"
-    ]
-  }
-}
