+++ src/soen343house/UserType.java
@@ -1,7 +1,6 @@
 package soen343house;
 
 public enum UserType {
-
 	PARENT,
 	CHILD,
 	GUEST,
